fix bdd migrations

+++ database/migrations/2021_05_26_083324_controle_etat_table.php
@@ -16,7 +16,7 @@ class ControleEtatTable extends Migration
         Schema::create('states_controls', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('id_vehicule')->unsigned();
-            $table->integer('id_employe')->unsigned();
+            $table->bigInteger('id_employe')->unsigned();
             $table->text('commentaire')->nullable();
             $table->timestamps();
